# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP Challenge API",
        "contact": {
            "name": "joaorezends",
            "url": "https://github.com/joaorezends"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/people": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get list of people",
                "description": "Returns list of people",
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get a person",
                "description": "Returns a person",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/shiporders": {
            "get": {
                "tags": [
                    "Shiporder"
                ],
                "summary": "Get list of shiporders",
                "description": "Returns list of shiporders",
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/shiporders/{id}": {
            "get": {
                "tags": [
                    "Shiporder"
                ],
                "summary": "Get a shiporder",
                "description": "Returns a shiporder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Shiporder id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}